rm v3 from package structure
