+++ hbc-example/pom.xml
@@ -24,7 +24,7 @@
   <dependencies>
     <dependency>
       <groupId>com.twitter</groupId>
-      <artifactId>hbc-twitter4j-v3</artifactId>
+      <artifactId>hbc-twitter4j</artifactId>
       <version>2.0.0-SNAPSHOT</version>
     </dependency>
     <dependency>
