+++ hbc-twitter4j/src/main/java/com/twitter/hbc/twitter4j/Twitter4jSitestreamClient.java
@@ -11,15 +11,15 @@
  * limitations under the License.
  **/
 
-package com.twitter.hbc.twitter4j.v3;
+package com.twitter.hbc.twitter4j;
 
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
 import com.twitter.hbc.core.Client;
-import com.twitter.hbc.twitter4j.v3.handler.SitestreamHandler;
-import com.twitter.hbc.twitter4j.v3.message.DisconnectMessage;
-import com.twitter.hbc.twitter4j.v3.message.StallWarningMessage;
-import com.twitter.hbc.twitter4j.v3.parser.JSONObjectParser;
+import com.twitter.hbc.twitter4j.handler.SitestreamHandler;
+import com.twitter.hbc.twitter4j.message.DisconnectMessage;
+import com.twitter.hbc.twitter4j.message.StallWarningMessage;
+import com.twitter.hbc.twitter4j.parser.JSONObjectParser;
 import twitter4j.*;
 
 import javax.annotation.Nullable;
